requirements.txt
# Face Recognition Attendance System - Python Dependencies
# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Computer Vision and Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# Face Recognition (ArcFace)
insightface==0.7.3
onnxruntime==1.16.3
scikit-learn==1.3.2

# Database
mysql-connector-python==8.2.0

# Data Processing
pandas==2.1.4
scipy==1.11.4
xlsxwriter==3.1.9

# Security and Encryption
cryptography>=3.0.0
bcrypt==4.1.2
PyJWT==2.8.0

# API and Web Services
requests==2.31.0
urllib3==2.0.7

# Additional ML Dependencies
torch==2.1.2
torchvision==0.16.2
onnx==1.15.0

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0

# Production Server (Optional)
gunicorn==21.2.0
waitress==2.1.2

# Monitoring and Logging
python-json-logger==2.0.7

# Environment and Configuration
python-decouple==3.8
